clean: move namespace
